snapenhance-backend/requirements.txt
flask
flask-cors
opencv-python
pillow
rembg
numpy
tensorflow
onnxruntime
pymongo
